kaggle/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
openai>=1.3.7
torch>=2.0.0
transformers>=4.36.0
sentence-transformers
faiss-cpu
nltk
scikit-learn
numpy>=1.24.0
pandas
accelerate
bitsandbytes
PyPDF2
pdfplumber
python-docx
python-pptx
pytesseract
Pillow
tiktoken
werkzeug
reportlab
pdf2image
python-dotenv
PyMuPDF
fpdf
